+++ src/main/java/org/example/fibonacci/FibonacciStrategy.java
@@ -1,5 +1,5 @@
 package org.example.fibonacci;
 
 public interface FibonacciStrategy<T extends Number> {
-    T solveFibonacci(long n);
+    T solveFibonacci(int n);
 }
